avoid warning
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@20163 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/jsbase/JSWrapper.h
@@ -788,7 +788,7 @@ AC_INFO<<"Registering function "<<theName;
         AC_INFO<<"Registering property "<<theName;
         JSPropertySpec mySpec;
         mySpec.name = theName;
-        mySpec.tinyid = theList.size()-100;
+        mySpec.tinyid = static_cast<int8>(theList.size()-100);
         mySpec.getter = theGetter;
         mySpec.setter = theSetter;
         uint8 myFlags = JSPROP_ENUMERATE|JSPROP_PERMANENT|JSPROP_SHARED;
